pyproject.toml
[project]
name = "interp-workbench"
version = "0.1.0"
description = "Interpretability workbench"
requires-python = "==3.12.*"
dependencies = [
    "fastapi>=0.115.6",
    "python-dotenv>=1.0.1",
    "nnsight>=0.5.12",
    # Required for .uv_sync on Modal image
    "modal>=0.65.66",
    "einops>=0.8.1",
    # Needed bc nnsight 0.5 requires 3.10 which does not have tomllib in stdlib
    "toml>=0.10.2",
    "dill>=0.4.0",
    "influxdb-client>=1.49.0",
    "uvicorn>=0.33.0",
    "torch>=2.9.0",
    "pytest>=9.0.1",
]

[project.optional-dependencies]
dev = [
    "ipykernel>=6.29.5",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.16",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["workbench"]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
